% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCM_plot.R
\name{eCM_plot}
\alias{eCM_plot}
\title{Enhanced Confusion Matrix Plot}
\usage{
eCM_plot(task, trained_model, splits, add_sums = TRUE, palette = "Green")
}
\arguments{
\item{task}{mlr3 task object specifying the task details}

\item{trained_model}{mlr3 trained learner (model) object obtained after training}

\item{splits}{mlr3 object defining data splits for train and test sets}

\item{add_sums}{logical, indicating whether total numbers should be displayed in the plot (default: TRUE)}

\item{palette}{character, the color palette for the confusion matrix (default: "Green")}
}
\value{
A confusion matrix plot visualizing sensitivity, specificity, PPV, and NPV
}
\description{
This function generates an enhanced confusion matrix plot using the CVMS package. The plot includes visualizations of sensitivity, specificity, positive predictive value (PPV), and negative predictive value (NPV).
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
