% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDecisionCurve.R
\name{eDecisionCurve}
\alias{eDecisionCurve}
\title{Decision Curve Plot}
\usage{
eDecisionCurve(task, trained_model, splits, seed = 246)
}
\arguments{
\item{task}{mlr3 task object specifying the task details}

\item{trained_model}{mlr3 trained learner (model) object obtained after training}

\item{splits}{mlr3 object defining data splits for train and test sets}

\item{seed}{numeric, seed for reproducibility (default: 246)}
}
\value{
An interactive decision curve plot
}
\description{
Decision curve analysis is a statistical method used in medical research to evaluate and compare the clinical utility of different diagnostic or predictive models. It assesses the net benefit of a model across a range of decision thresholds, aiding in the selection of the most informative and practical approach for guiding clinical decisions.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
