% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFairness.R
\name{eFairness}
\alias{eFairness}
\title{Enhanced Fairness Analysis}
\usage{
eFairness(task, trained_model, splits, target_variable, var_levels)
}
\arguments{
\item{task}{mlr3 binary classification task object specifying the task details}

\item{trained_model}{mlr3 trained learner (model) object obtained after training}

\item{splits}{mlr3 object defining data splits for train and test sets}

\item{target_variable}{character, the variable from the dataset used to test the model's performance against}

\item{var_levels}{list, defining the levels for the specified variable}
}
\value{
Model performance metrics for user-specified subgroups using Precision-Recall and ROC curves
}
\description{
This function generates Precision-Recall and ROC curves for sample subgroups, facilitating fairness analysis of a binary classification model.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
