% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eperformance.R
\name{eperformance}
\alias{eperformance}
\title{Enhanced Performance Evaluation}
\usage{
eperformance(task, trained_model, splits)
}
\arguments{
\item{task}{mlr3 binary classification task object specifying the task details}

\item{trained_model}{mlr3 trained learner (model) object obtained after training}

\item{splits}{mlr3 object defining data splits for train and test sets}
}
\value{
ROC and Precision-Recall curves with threshold information
}
\description{
This function generates Precision-Recall and ROC curves, including threshold information for binary classification models.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
