% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSHAP_plot.R
\name{eSHAP_plot}
\alias{eSHAP_plot}
\title{Enhanced SHAP Analysis for Binary Classification Models}
\usage{
eSHAP_plot(
  task,
  trained_model,
  splits,
  sample.size = 30,
  seed = 246,
  subset = 1
)
}
\arguments{
\item{task}{mlr3 task object for binary classification}

\item{trained_model}{mlr3 trained learner object}

\item{splits}{mlr3 object defining data splits for train and test sets}

\item{sample.size}{numeric, default to 30. The larger the value, the slower but more accurate the estimate of SHAP values}

\item{seed}{numeric, an integer for reproducibility. Default to 246}

\item{subset}{numeric, what percentage of the instances to use from 0 to 1 where 1 means all}
}
\value{
A list of two objects:
\enumerate{
\item An enhanced SHAP plot with user interactive elements,
\item A matrix of SHAP values
}
}
\description{
The SHAP plot for classification models is a visualization tool that uses the Shapley value, an approach from cooperative game theory, to compute feature contributions for single predictions. The Shapley value fairly distributes the difference of the instance’s prediction and the datasets average prediction among the features. This method is available from the iml package.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
