% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSHAP_plot_reg.R
\name{eSHAP_plot_reg}
\alias{eSHAP_plot_reg}
\title{Enhanced SHAP Analysis for Regression Models}
\usage{
eSHAP_plot_reg(
  task,
  trained_model,
  splits,
  sample.size = 30,
  seed = 246,
  subset = 1
)
}
\arguments{
\item{task}{mlr3 regression task object specifying the task details}

\item{trained_model}{mlr3 trained learner (model) object obtained after training}

\item{splits}{mlr3 object defining data splits for train and test sets}

\item{sample.size}{numeric, number of samples to calculate SHAP values (default: 30)}

\item{seed}{numeric, seed for reproducibility (default: 246)}

\item{subset}{numeric, proportion of the test set to use for visualization (default: 1)}
}
\value{
A list of two objects:
\enumerate{
\item An enhanced SHAP plot with user interactive elements,
\item A matrix of SHAP values
}
}
\description{
The SHAP plot for regression models is a visualization tool that uses the Shapley value, an approach from cooperative game theory, to compute feature contributions for single predictions. The Shapley value fairly distributes the difference of the instance’s prediction and the datasets average prediction among the features. This method is available from the iml package.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
