% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_anomalies.R
\name{explain_anomalies}
\alias{explain_anomalies}
\title{Explain anomalies found in the time series}
\usage{
explain_anomalies(object, burn, n_obs, seasonality, threshold = 3)
}
\description{
Explain anomalies found in the time series
}
\examples{
library(bsts)

y <- log(AirPassengers)
y[length(y) - 50] <- y[length(y) - 50] + 2
plot(y)

ss <- AddSemilocalLinearTrend(list(), y)
ss <- AddSeasonal(ss, y, nseasons = 12)
model <- bsts(y, state.specification = ss, niter = 5000)
pred <- predict(model, horizon = 12, burn = 1000)

par(mfrow = c(1,2))
plot(model)
plot(pred)

plot(model, "residuals")
plot(model, "prediction.errors")

explain_anomalies(
  object = model,
  burn = 1000,
  n_obs = length(y),
  seasonality = 12,
  threshold = 3
)

}
