% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_info.R
\name{forest_info}
\alias{forest_info}
\title{Function to extract tree information in the forest}
\usage{
forest_info(forest, ntree)
}
\arguments{
\item{forest}{name of the randomForest model}

\item{ntree}{number of trees in the forest}
}
\value{
retun a list of variable information corresponds to each tree
}
\description{
Function to extract tree information in the forest
}
\author{
Thiyanga Talagala
}
