% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedmanHstat.R
\name{friedmanHstat}
\alias{friedmanHstat}
\title{Fiedman's H-Statistic calculations}
\usage{
friedmanHstat(model, fulldf, subsetdf, allfeatures, trimfeatures,
  grid.resolution, classnames)
}
\arguments{
\item{model}{name of the fitted randomforest model}

\item{fulldf}{full data set used to fit the random forest}

\item{subsetdf}{subset dataset used to evaluate the fitted model}

\item{allfeatures}{vector of names of all features}

\item{trimfeatures}{vector of names of features that need to trim}

\item{grid.resolution}{number of levels to be considered in the interested features}

\item{classnames}{vector of names of class names}
}
\value{
data frame containing features and pedicted probabilities
}
\description{
Given two features calculate Friedman's H-Statistic between two features
}
\author{
Thiyanga Talagala
}
