% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{ice_cal}
\alias{ice_cal}
\title{Function to calculate Individual Conditional Expectation (ICE) plots}
\usage{
ice_cal(model, variable, fulldf, subsetdf, grid.resolution = 10,
  trim.outliers = FALSE)
}
\arguments{
\item{model}{name of the randomForest model}

\item{variable}{name of the interested feature}

\item{fulldf}{full data set used to fit the random forest}

\item{subsetdf}{subset of the dataset use to compute ice}

\item{grid.resolution}{number of levels to be considered in the interested feature}

\item{trim.outliers}{if TRUE range for the grid.resolution is computed trimming the outliers}
}
\value{
a dataframe including corresponding probability for each class
}
\description{
Function to create partial dependence plots given features and the training set
}
\author{
Thiyanga Talagala
}
