% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_ice.R
\name{meansd_ice}
\alias{meansd_ice}
\title{Function to calculate mean of standard deviation of ICE curves}
\usage{
meansd_ice(ice_data, variable, classlabelvec, subrw, grid.resolution)
}
\arguments{
\item{ice_data}{ice calculations for a specific variable(output of ice_cal)}

\item{variable}{name of the interested feature}

\item{classlabelvec}{vector of names corresponds to classlabels}

\item{subrw}{number of rows in the subset of the data frame we use to calculate ice curves in ice_cal function}

\item{grid.resolution}{gridresolution size used in ice_cal function}
}
\value{
a dataframe including corresponding probability for each class
}
\description{
ICE-based variable importance measure: mean of standard deviation of ICE curves
}
\author{
Thiyanga Talagala
}
