% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.explanation}
\alias{print.explanation}
\title{Print method for explanation objects}
\usage{
\method{print}{explanation}(explanation)
}
\arguments{
\item{explanation}{The object returned by \code{\link[=explain]{explain()}}}
}
\description{
Can be used to print the text components of the explanation in an
easy-to-read format.
}
