% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range01.R
\name{range01}
\alias{range01}
\title{Data Scale to 0 and 1}
\usage{
range01(x)
}
\arguments{
\item{x}{Vector or array of numbers to be normalized}
}
\value{
Normalized vector
}
\description{
Scale the data to the range of 0 to 1. It uses the Hampel filter to adjust outliers, followed by min-max normalization.
}
\examples{
\dontrun{
range01(seq(-10:1000))
}
}
