% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressmdl_eval.R
\name{regressmdl_eval}
\alias{regressmdl_eval}
\title{Regression Model Evaluation}
\usage{
regressmdl_eval(task, trained_model, splits)
}
\arguments{
\item{task}{mlr3 regression task object}

\item{trained_model}{mlr3 trained learner (model) object}

\item{splits}{mlr3 object defining data splits for train and test sets}
}
\value{
Data frame containing regression evaluation measures
}
\description{
Provides calculations of measures to evaluate regression models.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
