% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAPclust.R
\name{SHAPclust}
\alias{SHAPclust}
\title{clustered SHAP summary plot}
\usage{
SHAPclust(
  task,
  trained_model,
  splits,
  shap_Mean_wide,
  shap_Mean_long,
  num_of_clusters = 4,
  seed = 246,
  subset = 1
)
}
\arguments{
\item{task}{an mlr3 task for binary classification}

\item{trained_model}{an mlr3 trained learner object}

\item{splits}{an split object from mlr3 that defines train and test sets}

\item{shap_Mean_wide}{the data frame of SHAP values in wide format from eSHAP_plot.R}

\item{shap_Mean_long}{the data frame of SHAP values in long format from eSHAP_plot.R}

\item{num_of_clusters}{number of clusters to make based on SHAP values, default: 4}

\item{seed}{an integer for reproducibility, Default to 246}

\item{subset}{what percentage of the instances to use from 0 to 1 where 1 means all}
}
\description{
shap values are used to cluster data samples using Kmeans method that is to identify subgroups of individuals with specific patterns of feature contributions.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
