% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapFeaturePlot.R
\name{ShapFeaturePlot}
\alias{ShapFeaturePlot}
\title{SHAP values versus feature values}
\usage{
ShapFeaturePlot(shap_Mean_long)
}
\arguments{
\item{shap_Mean_long}{the data frame containing SHAP values in long format}
}
\value{
interactive plot of SHAP values in association with feature values
}
\description{
SHAP values in association with feature values
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
