% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapPartialPlot.R
\name{shapPartialPlot}
\alias{shapPartialPlot}
\title{SHAP Partial Plot}
\usage{
shapPartialPlot(shap_Mean_long)
}
\arguments{
\item{shap_Mean_long}{data frame containing SHAP values in long format}
}
\value{
Interactive partial dependence plot
}
\description{
Generates an interactive partial dependence plot based on SHAP values, visualizing the marginal effect of one or two features on the predicted outcome of a machine learning model.
}
\examples{
\dontrun{
# To see detailed examples, refer to the example tutorials in the package vignettes.
}
}
