% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twowaypartial.R
\name{twowayinteraction}
\alias{twowayinteraction}
\title{Two-way interaction plots}
\usage{
twowayinteraction(model, variable1, variable2, fulldf, subsetdf,
  grid.resolution = 10, trim1 = FALSE, trim2 = FALSE)
}
\arguments{
\item{model}{name of the fitted randomforest model}

\item{variable1}{name of the first variable}

\item{variable2}{name of the second variable}

\item{fulldf}{full data set used to fit the random forest}

\item{subsetdf}{subset dataset used to evaluate the fitted model}

\item{grid.resolution}{number of levels to be considered in the interested features}

\item{trim1}{if TRUE remove outliers of the first variable}

\item{trim2}{if TRUE remove outliers of the second variable}
}
\value{
data frame containing features and pedicted probabilities
}
\description{
Given two features create partial two way interaction contourplots
}
\author{
Thiyanga Talagala
}
