% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_pdp.R
\name{sd_pdp}
\alias{sd_pdp}
\title{Function to calculate standard deviation of partial dependence curves}
\usage{
sd_pdp(icedata, variable, classlabelvec)
}
\arguments{
\item{icedata}{ice calculations for a specific variable(output of icecal)}

\item{variable}{name of the interested feature}

\item{classlabelvec}{vector of names corresponds to classlabels}
}
\value{
a dataframe including corresponding probability for each class
}
\description{
PDP-based variable importance measure: standard deviation of partial dependence curves
}
\author{
Thiyanga Talagala
}
