% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAPclust.R
\name{SHAPclust}
\alias{SHAPclust}
\title{Clustered SHAP Summary Plot}
\usage{
SHAPclust(
  task,
  trained_model,
  splits,
  shap_Mean_wide,
  shap_Mean_long,
  num_of_clusters = 4,
  seed = 246,
  subset = 1
)
}
\arguments{
\item{task}{an mlr3 task for binary classification}

\item{trained_model}{an mlr3 trained learner object}

\item{splits}{an mlr3 object defining data splits for train and test sets}

\item{shap_Mean_wide}{the data frame of SHAP values in wide format from eSHAP_plot.R}

\item{shap_Mean_long}{the data frame of SHAP values in long format from eSHAP_plot.R}

\item{num_of_clusters}{number of clusters to make based on SHAP values, default: 4}

\item{seed}{an integer for reproducibility, Default to 246}

\item{subset}{what percentage of the instances to use from 0 to 1 where 1 means all}
}
\description{
SHAP values are used to cluster data samples using the Kmeans method to identify subgroups of individuals with specific patterns of feature contributions.
}
\examples{
\donttest{
library("explainer")
seed <- 246
set.seed(seed)
# Install and load necessary packages
if (!requireNamespace("mlbench", quietly = TRUE)) {
  install.packages("mlbench")
  library(mlbench)
}
if (!requireNamespace("mlr3learners", quietly = TRUE)) {
  install.packages("mlr3learners")
  library(mlr3learners)
}
if (!requireNamespace("ranger", quietly = TRUE)) {
  install.packages("ranger")
  library(ranger)
}
# Load BreastCancer dataset
utils::data("BreastCancer", package = "mlbench")
target_col <- "Class"
positive_class <- "malignant"
mydata <- BreastCancer[, -1]
mydata <- na.omit(mydata)
sex <- sample(
  c("Male", "Female"),
  size = nrow(mydata),
  replace = TRUE
)
mydata$age <- as.numeric(sample(
  seq(18,60),
  size = nrow(mydata),
  replace = TRUE
))
mydata$sex <- factor(
  sex,
  levels = c("Male", "Female"),
  labels = c(1, 0)
)
maintask <- mlr3::TaskClassif$new(
  id = "my_classification_task",
  backend = mydata,
  target = target_col,
  positive = positive_class
)
splits <- mlr3::partition(maintask)
mylrn <- mlr3::lrn(
  "classif.ranger",
  predict_type = "prob"
)
mylrn$train(maintask, splits$train)
SHAP_output <- eSHAP_plot(
  task = maintask,
  trained_model = mylrn,
  splits = splits,
  sample.size = 30,
  seed = seed,
  subset = 0.8
)
SHAP_plot_clusters <- SHAPclust(
  task = maintask,
  trained_model = mylrn,
  splits = splits,
  shap_Mean_wide = shap_Mean_wide,
  shap_Mean_long = shap_Mean_long,
  num_of_clusters = 4,
  seed = seed,
  subset = 0.8
)
}
}
